+++ src/bitboardhelper.cpp
@@ -35,7 +35,7 @@ int BBHelper::square_to_bit(int square) {
     return GRID_NUM_SQUARES - square - 1;
 }
 
-int BBHelper::bit_to_square(int bit) {
+uint8_t BBHelper::bit_to_square(uint8_t bit) {
     return GRID_NUM_SQUARES - bit - 1;
 }
